refactor(api-reference): move createEmptySpecification to libs/openapi and update imports
